requirements.txt
celery
channels
daphne
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
Pillow
prompt-toolkit
psycopg
psycopg2-binary
python-decouple
